backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
spotipy==2.23.0
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0